👩‍🌾 Clear all Windows warnings (#206)

+++ src/OrbitViewController.cc
@@ -41,8 +41,8 @@ class ignition::rendering::OrbitViewControllerPrivate
 using namespace ignition;
 using namespace rendering;
 
-static const float PITCH_LIMIT_LOW = -IGN_PI*0.5 + 0.001;
-static const float PITCH_LIMIT_HIGH = IGN_PI*0.5 - 0.001;
+static const float PITCH_LIMIT_LOW = -static_cast<float>(IGN_PI)*0.5f + 0.001f;
+static const float PITCH_LIMIT_HIGH = static_cast<float>(IGN_PI)*0.5f - 0.001f;
 
 //////////////////////////////////////////////////
 OrbitViewController::OrbitViewController()
